separate subgraph interfaces
