+++ packages/core/src/interfaces/ubi/beneficiary.ts
@@ -86,12 +86,3 @@ export interface BeneficiaryCreationAttributes {
     tx: string;
     txAt: Date;
 }
-
-export interface BeneficiarySubgraph {
-    claims: number;
-    community: {
-        id: string;
-    };
-    lastClaimAt: number;
-    preLastClaimAt: number;
-}
