+++ packages/core/src/subgraph/index.ts
@@ -1,4 +1,5 @@
 import * as config from './config';
 import * as queries from './queries';
+import * as interfaces from './interfaces';
 
-export { queries, config };
+export { queries, config, interfaces };
