+++ packages/core/src/subgraph/interfaces/beneficiary.ts
@@ -1,0 +1,8 @@
+export interface BeneficiarySubgraph {
+    claims: number;
+    community: {
+        id: string;
+    };
+    lastClaimAt: number;
+    preLastClaimAt: number;
+}
