+++ packages/core/src/subgraph/interfaces/index.ts
@@ -1,0 +1,3 @@
+import * as beneficiary from './beneficiary';
+
+export { beneficiary };
